requirements.txt
keyboard
RPi.GPIO